avoid polluting URL hash and ignore stale block data updates
