+++ modules/periodo-app/src/backends/components/BackendHome.js
@@ -24,7 +24,6 @@ section = Filter periods
 [Facets]
 type = facets
 section = Filter periods
-flex = true
 height = 200
 
 [SpatialCoverage]
@@ -118,7 +117,6 @@ module.exports = class BackendHome extends React.Component {
               dataset,
               gazetteers,
               layout,
-              totalCount: dataset.periods.length,
               selectedPeriod: dataset.periodByID(periodID),
               blockOpts,
               onBlockOptsChange: updatedOpts => {
