+++ modules/periodo-app/src/layouts/authorities/Facets/index.js
@@ -4,10 +4,10 @@ const h = require('react-hyperscript')
     , R = require('ramda')
     , React = require('react')
     , tags = require('language-tags')
-    , { Flex, Box, Link, Label, HelpText } = require('periodo-ui')
+    , { Label, HelpText } = require('periodo-ui')
     , { period: { authorityOf }} = require('periodo-utils')
     , { shallowEqualObjects } = require('shallow-equal')
-    , styled = require('styled-components').default
+    , AspectTable = require('./AspectTable')
     , FacetCalculator = require('./FacetCalculator')
 
 const languageDescription = R.memoize(tag => {
@@ -41,183 +41,6 @@ const aspects = {
     getter: period => period.spatialCoverageDescription || null,
     render: identityWithDefault('(no value)'),
   },
-}
-
-const Table = Box.extend([], {
-  overflowY: 'scroll',
-  '& td': {
-    padding: '2px 5px',
-  },
-  '& td:first-of-type': {
-    color: '#999',
-  },
-  '& td:last-of-type': {
-    width: '100%',
-  },
-})
-
-function withoutValue(val, set) {
-  const newSet = new Set(set)
-  newSet.delete(val)
-  return newSet;
-}
-
-function withValue(val, set) {
-  const newSet = new Set(set)
-  newSet.add(val)
-  return newSet;
-}
-
-const AspectContainer = styled(Flex)`
-&:not(:last-of-type) {
-  margin-right: 16px;
-}
-`
-
-class AspectTable extends React.Component {
-  constructor() {
-    super()
-  }
-
-  shouldComponentUpdate(prevProps) {
-    return (
-      prevProps.counts !== this.props.counts
-    )
-  }
-
-  render() {
-    const { aspect, aspectID, opts, updateOpts, counts, hidden } = this.props
-        , { label, flexBasis } = aspect
-        , render = aspect.render || R.identity
-        , height = parseInt(opts.height || '256')
-        , selected = new Set(R.path([ 'selected', aspectID ], opts) || [])
-
-    if (hidden.has(aspectID)) {
-      return null
-    }
-
-    const selectedRows = []
-        , unselectedRows = []
-
-    if (counts) {
-      counts.forEach(([ value, count, label ]) => {
-        const isSelected = selected.has(value)
-
-        const el = (
-          h('tr', [
-            h('td', count),
-            h('td', [
-              h(Link, {
-                href: '',
-                color: `blue.${ isSelected ? 8 : 4 }`,
-                onClick: e => {
-                  e.preventDefault();
-                  updateOpts(R.pipe(
-                    R.over(
-                      R.lensPath([ 'selected', aspectID ]),
-                      () => [ ...(isSelected
-                        ? withoutValue(value, selected)
-                        : withValue(value, selected)) ]),
-                    R.ifElse(
-                      val => val.selected[aspectID].length,
-                      R.identity,
-                      R.dissocPath([ 'selected', aspectID ])),
-                    R.ifElse(
-                      val => R.isEmpty(val.selected),
-                      R.dissoc('selected'),
-                      R.identity)
-                  ), true)
-
-                },
-              }, render(label === undefined ? value : label)),
-            ]),
-          ])
-        )
-
-        if (isSelected) {
-          selectedRows.push(el)
-        } else {
-          unselectedRows.push(el)
-        }
-      })
-    }
-
-    return (
-      h(AspectContainer, {
-        style: {
-          flex: 1,
-          flexBasis,
-        },
-        flexDirection: 'column',
-        height,
-      }, [
-        h(Flex, {
-          justifyContent: 'space-between',
-          alignItems: 'center',
-          bg: 'gray.1',
-          p: 2,
-          fontWeight: 'bold',
-          fontSize: 1,
-        }, [
-          h('span', label),
-
-          h('span', {}, selected.size === 0 ? null : (
-            h('a', {
-              href: '',
-              style: {
-                color: 'red',
-                textDecoration: 'none',
-                fontSize: '14px',
-                fontWeight: 100,
-              },
-              onClick: e => {
-                e.preventDefault();
-                updateOpts(R.dissocPath([ 'selected', aspectID ]), true)
-              },
-            }, 'Clear')
-          )),
-        ]),
-
-        h('div', {
-          style: {
-            height:'100%',
-            overflowY: 'scroll',
-            overscrollBehaviorY: 'contain',
-          },
-        }, counts == null ? (
-          h('div', {
-            style: {
-              textAlign: 'center',
-              padding: '1em',
-            },
-          }, '. . .')
-        ): [
-          selectedRows.length === 0 ? null : (
-            h(Table, {
-              is: 'table',
-              className: 'selected',
-              px: 1,
-              py: 1,
-              width: '100%',
-            }, [
-              h('tbody', selectedRows),
-            ])
-          ),
-
-          unselectedRows.length === 0 ? null : (
-            h(Table, {
-              is: 'table',
-              px: 1,
-              width: '100%',
-              bg: 'gray.1',
-            }, [
-              h('tbody', unselectedRows),
-            ])
-          ),
-        ]),
-      ])
-    )
-  }
 }
 
 class Facets extends React.Component {
@@ -282,44 +105,43 @@ class Facets extends React.Component {
     const { opts, data, updateOpts, dataset } = this.props
         , { countsByAspect } = this.state
 
-    const style = {}
-
-    if (opts.flex) {
-      style.display = 'flex'
-    }
-
-    const hidden = new Set(opts.hidden|| [])
-
-    if (hidden.size === Object.keys(aspects).length) {
-      return null
-    }
+    const hiddenAspects = new Set(
+      opts.hiddenAspects ? opts.hiddenAspects.split(',') : []
+    )
 
     const aspectsShown = Object.entries(aspects)
-      .filter(([ k ]) => ! hidden.has(k))
+      .filter(([ k ]) => ! hiddenAspects.has(k))
+
+    const aspectsDescription = aspectsShown
       .map(([ , v ]) => v.label.toLowerCase())
       .join(', ')
       .replace(/,([^,]*)$/, ' or$1')
 
+    const aspectProportions = opts.aspectProportions
+      ? opts.aspectProportions.split(',')
+      : []
+
     return (h('div'), [
       h(Label, { key: 'label' },
-        `By ${aspectsShown}`),
+        `By ${aspectsDescription}`),
 
       h(HelpText, { key: 'help' },
-        `Show periods having the selected ${aspectsShown}`),
+        `Show periods having the selected ${aspectsDescription}`),
 
       h('div', {
-        style,
+        style: { display: 'flex' },
         key: 'aspect-table',
-      }, Object.entries(aspects).map(([ key, aspect ]) =>
+      }, aspectsShown.map(([ key, aspect ], i) =>
         h(AspectTable, {
           key,
           data,
-          flex: opts.flex,
           rawDataset: dataset.raw,
-          aspect,
+          aspect: {
+            ...aspect,
+            flexBasis: aspectProportions[i] || aspect.flexBasis,
+          },
           aspectID: key,
           counts: countsByAspect[key],
-          hidden,
           opts,
           updateOpts,
         })
